Travis: Run PHPCS in two parts, plugin files and test files (needs PHPCS 3.3.0)
The plugin files in the ACF analysis plugin have to be compatible with PHP 5.2.
However, the test files have been set up to work only on PHP 5.6+.

The YoastCS base ruleset checks files for PHP compatibility against PHP 5.2+.

As of PHPCS 3.3.0, it is now possible to overrule the config directive which instructs the PHPCompatibility library against which PHP version to check, from the ruleset.

So by running PHPCS over all files in two batches - plugin files and test files - we prevent false positives for the unit tests not being compatible with PHP 5.2-5.5, while still being able to check their compatibility.

Also:
* Enabled the "ignore warnings on exit" flag to allow the build to pass while there is still some technical debt.